### 2026-03-31 #198 @editorialbot
[PRE REVIEW]: StreamSampling.jl: Efficient Sampling from Data Streams in Julia
**Submitting author:** @ameligrana (<a href="http://orcid.org/0009-0001-0410-5964">Adriano Meligrana</a>)
**Repository:** https://github.com/JuliaDynamics/StreamSampling.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
**Managing EiC:** Ludovic Räss

Thanks for submitting your paper to JCON @ameligrana. **Currently, there isn't a JCON editor assigned** to your paper.
