### 2026-03-31 4e1a508
Some formatting updates to the paper (#140)
