+++ paper/ref.bib
@@ -37,7 +37,7 @@ number = {2},
 pages = {750-762},
 year = {2007},
 issn = {0167-9473},
-doi = {https://doi.org/10.1016/j.csda.2007.03.010},
+doi = {10.1016/j.csda.2007.03.010},
 url = {https://www.sciencedirect.com/science/article/pii/S0167947307001089},
 author = {Byung-Hoon Park and George Ostrouchov and Nagiza F. Samatova},
 keywords = {Data stream mining, Random sampling with replacement, Reservoir sampling},
@@ -51,7 +51,7 @@ number = {5},
 pages = {181-185},
 year = {2006},
 issn = {0020-0190},
-doi = {https://doi.org/10.1016/j.ipl.2005.11.003},
+doi = {10.1016/j.ipl.2005.11.003},
 url = {https://www.sciencedirect.com/science/article/pii/S002001900500298X},
 author = {Pavlos S. Efraimidis and Paul G. Spirakis},
 keywords = {Weighted random sampling, Reservoir sampling, Randomized algorithms, Data streams, Parallel algorithms},
@@ -91,7 +91,6 @@ numpages = {10}
   booktitle = {Proceedings of The 24th International Conference on Artificial Intelligence and Statistics},
   pages = {3628--3636},
   year = {2021},
-  editor = {Banerjee, Arindam and Fukumizu, Kenji},
   volume = {130},
   series = {Proceedings of Machine Learning Research},
   publisher = {PMLR},
@@ -206,14 +205,13 @@ journal = {Journal of Open Source Software}
 @InProceedings{lee2010stratified,
 author="Al-Kateb, Mohammed
 and Lee, Byung Suk",
-editor="Gertz, Michael
-and Lud{\"a}scher, Bertram",
 title="Stratified Reservoir Sampling over Heterogeneous Data Streams",
 booktitle="Scientific and Statistical Database Management",
 year="2010",
 publisher="Springer Berlin Heidelberg",
 pages="621--639",
-isbn="978-3-642-13818-8"
+isbn="978-3-642-13818-8",
+doi="10.1007/978-3-642-13818-8_42",
 }
 
 @article{braverman2012optimal,
@@ -225,7 +223,7 @@ pages = {260-272},
 year = {2012},
 note = {JCSS Knowledge Representation and Reasoning},
 issn = {0022-0000},
-doi = {https://doi.org/10.1016/j.jcss.2011.04.004},
+doi = {10.1016/j.jcss.2011.04.004},
 url = {https://www.sciencedirect.com/science/article/pii/S0022000011000493},
 author = {Vladimir Braverman and Rafail Ostrovsky and Carlo Zaniolo},
 keywords = {Data streams, Random sampling, Sliding windows},
