### 2026-03-31 7abde2a
Fix DOI formatting in reference.bib
