### 2026-03-31 c4ec17a
Update references to respect editorialbot constraints
