+++ paper/ref.bib
@@ -63,7 +63,6 @@ title = {Weighted Reservoir Sampling With Replacement from Data Streams},
 year = {2026},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
-doi = {10.1145/3774904.3792966},
 booktitle = {Proceedings of the ACM on Web Conference 2026},
 note={Accepted for publication. Currently available as arXiv preprint arXiv:2403.20256}
 }
