+++ paper/ref.bib
@@ -210,7 +210,7 @@ year="2010",
 publisher="Springer Berlin Heidelberg",
 pages="621--639",
 isbn="978-3-642-13818-8",
-doi="10.1007/978-3-642-13818-8\_42",
+doi="10.1007/978-3-642-13818-8",
 }
 
 @article{braverman2012optimal,
