### 2026-03-31 8536d5b
Update references to respect editorialbot constraints (#141)
* Update references to respect editorialbot constraints

* Update ref.bib
