### 2026-04-02 #198 @luraess
Hi all, given we have 2 reviewers willing to help out here we can proceed to the revision phase. This will close this PR and open another PR for the revision where we can continue to iterate. Note that we have Issue-style revision processs as JOSS does.
