### 2026-04-02 #202 @editorialbot
[REVIEW]: StreamSampling.jl: Efficient Sampling from Data Streams in Julia
**Submitting author:** @ameligrana (<a href="http://orcid.org/0009-0001-0410-5964">Adriano Meligrana</a>)
**Repository:** https://github.com/JuliaDynamics/StreamSampling.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
