### 2026-04-02 #202 @jbytecode
@ameligrana - It would be nice to have installation instructions in the README.md file of repository. I know, it has been mentioned in the documentation but an entry in the README like

## Installation

```julia
julia> using Pkg
julia> Pkg.add("StreamSampling")
```

and

```julia
julia> ]
(@v1.xx) pkg> add https://github.com/JuliaDynamics/StreamSampling.jl.git
```
with some descriptive text would be perfect.

Here is an example: https://github.com/jbytecode/OperationsResearchModels.jl/blob/main/README.md
