### 2026-04-02 #202 @luraess
Thanks for getting started here! As general recommendation I would suggest to post issues related to the repo/code directly on the related code repository (linking to this PR for book-keeping), and possibly posting feedback about the paper in this thread. Proceeding as such may help for overview and clarity 🙏
