### 2026-04-02 #202 @jbytecode
@luraess - Sure, I'll open issues for future suggestions on code and documentation and the other stuff about the repo.
