### 2026-04-02 Tortar/IteratorSampling.jl#143 @jbytecode
Community guidelines
Due to the submission: https://github.com/JuliaCon/proceedings-review/issues/202

In the review check list, there is an item that stating

In your readme.md, there is a paragraph about that but I think this section would be improved by stressing how to contribute, repoting format (I think you have 3 kind of issues), and how to seek support (are you hanging in Julia slack channel maybe, or are you active in Julia Discourse?, etc)
