### 2026-04-03 5b78cfb
Add installation instructions to README (#144)
