### 2026-04-03 Tortar/IteratorSampling.jl#142 @ameligrana
thank you @jbytecode, added the instructions in the more or less same format you suggested
