### 2026-04-03 22259ca
Enhance README with support and contribution details (#145)
