+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -12,6 +12,3 @@ A clear and concise description of what the bug is.
 
 **To Reproduce**
 Steps to reproduce the behavior.
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
