+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -7,11 +7,5 @@ assignees: ''
 
 ---
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is.
-
 **Describe the solution you'd like**
 A clear and concise description of what you want to happen.
-
-**Additional context**
-Add any other context about the feature request here.
