### 2026-04-03 Tortar/IteratorSampling.jl#143 @ameligrana
To try to address the concerns I added a new section for support related inquiries which points to the Discussions section of the repo, which I enabled for this. I also added an issue template for bugs and feature requests. Let me know if you think this is enough!
