### 2026-04-04 Tortar/IteratorSampling.jl#148 @riteshbhirud
Paper: Missing "Research Impact Statement" section
Per JOSS [review criteria](https://joss.readthedocs.io/en/latest/review_checklist.html), the paper needs a section providing evidence of realized impact or credible near-term significance.

The paper already contains strong material that could serve this purpose:
- Benchmark results (Section 5) demonstrating performance and memory improvements over `StatsBase.sample`
- The on-disk sampling application (Section 6) with 100 GB datasets
- The published research on AlgWRSWRSKIP (reference [9])

However, this isn't framed as a "Research Impact" section. I'd suggest, if possible:
Adding a brief dedicated subsection explicitly framing the above as evidence of near-term scholarly significance,

If no external adoption evidence exists yet, the benchmarks and the accepted publication are likely sufficient, but the framing should be made explicit.
