### 2026-04-04 Tortar/IteratorSampling.jl#149 @riteshbhirud
Paper: Missing required "AI Usage Disclosure" section
Per JOSS review guidelines, the paper must include an "AI usage disclosure" section that transparently discloses any use of generative AI in:
- Software creation or development
- Documentation writing
- Paper authoring

If no AI tools were used, this should be explicitly stated.
