### 2026-04-04 #202 @riteshbhirud
Review summary: StreamSampling.jl is a well-engineered, comprehensive Julia package that fills a genuine gap in the ecosystem. The software quality is high, the code is well-structured, thoroughly tested (including statistical correctness tests), documented, and follows good open-source practices. The paper is clearly written and makes a compelling case for the package. The following are some action items for the author:

- **Software design section**: Opened [issue](https://github.com/JuliaDynamics/StreamSampling.jl/issues/147). The current "Sampling API" section covers API usage but doesn't discuss the architectural trade-offs and design decisions behind the software. Suggest adding a dedicated section.

- **Research impact statement section**: Opened [issue](https://github.com/JuliaDynamics/StreamSampling.jl/issues/148). The evidence is largely present in the paper (benchmarks, on-disk application, accepted publication on AlgWRSWRSKIP) but isn't framed as a research impact statement.

- **AI usage disclosure section**: Opened [issue](https://github.com/JuliaDynamics/StreamSampling.jl/issues/149) This required section is currently missing from the paper.
