### 2026-04-11 #202 @riteshbhirud
@luraess  This library appears solid based on my testing and usage. The paper is well written, and with the three suggestions mentioned above which are the addition of the corresponding sections in accordance with the official JOSS guidelines, I would recommend it for acceptance. Great work!
