### 2026-05-04 5614301
relax compat restrictions on DataStructures (#150)
Fixes issues with Agents.jl and other packages
