+++ v0.8.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"category":"section","location":"api/#API","page":"API","text":"","title":"API"},{"category":"section","location":"api/#Types","page":"API","text":"","title":"Types"},{"category":"section","location":"api/#Methods","page":"API","text":"","title":"Methods"},{"category":"section","location":"api/#Algorithms","page":"API","text":"","title":"Algorithms"},{"category":"type","location":"api/#StreamSampling.ReservoirSampler","page":"API","text":"ReservoirSampler{T}([rng], method = AlgRSWRSKIP())\nReservoirSampler{T}([rng], n::Int, method = AlgL(); ordered = false)\n\nInitializes a reservoir sampler with elements of type T.\n\nThe first signature represents a sample where only a single element is collected. If ordered is true, the sampled values can be retrived in the order they were collected using ordvalue.\n\nLook at the Sampling Algorithms section for the supported methods. \n\n\n\n\n\n","title":"StreamSampling.ReservoirSampler"},{"category":"type","location":"api/#StreamSampling.SequentialSampler","page":"API","text":"SequentialSampler{T}([rng], iter, n, [N], method = AlgD())\n\nInitializes a sequential sampler, which can then be iterated over to return the sampling elements of the iterable iter which is assumed to have a eltype of T. The methods implemented in SequentialSampler require the knowledge of the total number of elements in the stream N, if not provided it is assumed to be available by calling length(iter).\n\n\n\nSequentialSampler{T}([rng], iter, wfunc, n, W, method = AlgORDWSWR())\n\nInitializes a weigthed sequential sampler, which can then be iterated over to return the sampling elements of the iterable iter which is assumed to have a eltype of T. The methods implemented in SequentialSampler for weighted streams require the knowledge of the total weight of the stream W and a weight function wfunc specifying how to map an element to its weight.\n\n\n\nSequentialSampler([rng], n::Integer, N::Integer, method = AlgD())\n\nInitializes a sequential sampler, which can then be iterated over to return n ordered indices between 1 and N, respecting the sampling scheme of the selected method, which can be AlgD(), AlgHiddenShuffle() or AlgORDSWR().\n\n\n\n\n\n","title":"StreamSampling.SequentialSampler"},{"category":"function","location":"api/#StatsAPI.fit!","page":"API","text":"fit!(rs::AbstractReservoirSampler, el)\nfit!(rs::AbstractReservoirSampler, el, w)\n\nUpdates the reservoir sample by taking into account the element passed. If the sampling is weighted also the weight of the elements needs to be passed.\n\n\n\n\n\n","title":"StatsAPI.fit!"},{"category":"function","location":"api/#Base.merge!","page":"API","text":"Base.merge!(rs::AbstractReservoirSampler...)\n\nUpdates the first reservoir sampler by merging its value with the values of the other samplers. The number of elements after merging will be the minimum number of elements in the merged reservoirs.\n\n\n\n\n\n","title":"Base.merge!"},{"category":"function","location":"api/#Base.merge","page":"API","text":"Base.merge(rs::AbstractReservoirSampler...)\n\nCreates a new reservoir sampler by merging the values of the samplers passed. The number of elements in the new sampler will be the minimum number of elements in the merged reservoirs.\n\n\n\n\n\n","title":"Base.merge"},{"category":"function","location":"api/#StreamSampling.combine","page":"API","text":"combine([rng], samples::AbstractArray, weights::AbstractArray)\n\nCombines different sequential samples in a single one. The number of  elements in the new sampler will be the minimum number of elements in the sa
[…]
