+++ paper/ref.bib
@@ -59,12 +59,19 @@ keywords = {Weighted random sampling, Reservoir sampling, Randomized algorithms,
 
 @inproceedings{meligrana2026weighted,
 author = {Meligrana, Adriano and Fazzone, Adriano},
-title = {Weighted Reservoir Sampling With Replacement from Data Streams},
+title = {Weighted Reservoir Sampling with Replacement from Data Streams},
 year = {2026},
+isbn = {9798400723070},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
-booktitle = {Proceedings of the ACM on Web Conference 2026},
-note={Accepted for publication. Currently available as arXiv preprint arXiv:2403.20256}
+url = {https://doi.org/10.1145/3774904.3792966},
+doi = {10.1145/3774904.3792966},
+booktitle = {Proceedings of the ACM Web Conference 2026},
+pages = {8789–8792},
+numpages = {4},
+keywords = {reservoir sampling, random sampling, weighted sampling, data streams},
+location = {United Arab Emirates},
+series = {WWW '26}
 }
 
 @article{vitter1987efficient,
@@ -85,16 +92,16 @@ numpages = {10}
 }
 
 @InProceedings{shekelyan2021sequential,
-  title = {Sequential Random Sampling Revisited: Hidden Shuffle Method},
-  author = {Shekelyan, Michael and Cormode, Graham},
-  booktitle = {Proceedings of The 24th International Conference on Artificial Intelligence and Statistics},
-  pages = {3628--3636},
-  year = {2021},
-  volume = {130},
-  series = {Proceedings of Machine Learning Research},
-  publisher = {PMLR},
-  pdf = {http://proceedings.mlr.press/v130/shekelyan21a/shekelyan21a.pdf},
-  url = {https://proceedings.mlr.press/v130/shekelyan21a.html},
+title = {Sequential Random Sampling Revisited: Hidden Shuffle Method},
+author = {Shekelyan, Michael and Cormode, Graham},
+booktitle = {Proceedings of The 24th International Conference on Artificial Intelligence and Statistics},
+pages = {3628--3636},
+year = {2021},
+volume = {130},
+series = {Proceedings of Machine Learning Research},
+publisher = {PMLR},
+pdf = {http://proceedings.mlr.press/v130/shekelyan21a/shekelyan21a.pdf},
+url = {https://proceedings.mlr.press/v130/shekelyan21a.html},
 }
 
 @article{bentley1980generating,
@@ -115,52 +122,52 @@ numpages = {6}
 }
 
 @article{startek2016asymptotically,
-  author = {Michal Startek},
-  title = {An asymptotically optimal, online algorithm for weighted random sampling with replacement},
-  journal      = {CoRR},
-  volume       = {abs/1611.00532},
-  year         = {2016},
-  url          = {http://arxiv.org/abs/1611.00532},
-  eprinttype    = {arXiv},
-  eprint       = {1611.00532},
-  biburl       = {https://dblp.org/rec/journals/corr/Startek16.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
+author = {Michal Startek},
+title = {An asymptotically optimal, online algorithm for weighted random sampling with replacement},
+journal = {CoRR},
+volume = {abs/1611.00532},
+year = {2016},
+url = {http://arxiv.org/abs/1611.00532},
+eprinttype = {arXiv},
+eprint = {1611.00532},
+biburl = {https://dblp.org/rec/journals/corr/Startek16.bib},
+bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
[…]
