### 2026-05-11 a0fada8
Bump to 0.8.2
