+++ paper/paper.tex
@@ -2,7 +2,17 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 \usepackage{balance}
+
 \usepackage{hyperref}
+\hypersetup{
+    colorlinks=true,
+    linkcolor=blue,
+    citecolor=blue,
+    urlcolor=blue,
+    filecolor=blue,
+    pdfborder={0 0 0},
+    breaklinks=true
+}
 
 \begin{document}
 
