### 2026-05-11 #202 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
